Don't delete archive in dryrun mode!
Former-commit-id: 4e9144f95cbce0b9b1377ed1cdfeb40912e60812 [formerly 0d44d222694254eae6f18d9e90a804d3d987cc64 [formerly b0134c4efffce9179814ceb2dd64488ce08f3329]]
Former-commit-id: ff582409970d8df5b13677c7c7a3a07b7044eba3 [formerly 51803ecc5d4b3bfcb5c7488a86cb3168bf2f7fc4]
Former-commit-id: 5dc8a3e2a7c23501f188721427e19363350d90ac

+++ scripts/Create-DraftRelease.ps1
@@ -353,7 +353,7 @@ if ($dryrun -or $buildonly -or $release.assets_url -imatch '^https:') {
         "`nYou should delete this bundle unless you are very sure that it was built with the correct certificate!"
         $deleteBundle = Read-Host "WARNING: There is already an existing UWP release package! Delete (recommended)? [Y/N]"
         $deleteBundle = -Not ( $deleteBundle -imatch 'n' )
-        If ($deleteBundle) {
+        If ((-Not $dryrun) -and $deleteBundle) {
           rm $ReleaseBundle
         }
     }
